fix: accumulated fixes for Phase 2 pipeline
- launch_pipeline.sh: set PYTHONPATH for all SLURM stages
- generate.py: clone output_tokens before run_with_cache to avoid in-place mutation
- logit_lens.py: cast residual to unembed dtype before matmul
- sae_analysis.py: load SAE directly to cuda
- run_logit_lens.py: keep ln_final/unembed on GPU; use MAX_NEW_TOKENS (4096) instead of 256
- visualize.py: layer recommendation uses SAE signal only, drops logit-lens rank

+++ src/logit_lens.py
@@ -12,7 +12,7 @@ def project_to_logits(
 ) -> torch.Tensor:
     """Project residual stream to logit space via final LayerNorm + unembedding."""
     normed = ln_final(residual)
-    return normed @ unembed_matrix
+    return normed.to(unembed_matrix.dtype) @ unembed_matrix
 
 
 def compute_token_divergence(
